+++ CHANGELOG.md
@@ -17,6 +17,6 @@
 ## 0.0.5
 
 - Add utility for pre/post processing (no testing in alpha)
-- Updated support for use postcode for county, town
+- Updated support for use postcode for county, town, and more.
 - Updated support for use locality for town
 - Updated readme for devs